Update unit test
Also, get diffscan_example back
