+++ setup.py
@@ -1,6 +1,6 @@
 from distutils.core import setup
 
 setup(name='SGMPy',
-      version='0.1.0 alpha',
+      version='0.1.1',
       packages=['scripts', 'data'],
       )
